requirements.txt
fastapi
uvicorn[standard]
tweepy
python-dotenv
